Split setup routines for multi-index tests out.
This moves the setup code for the combined search integration tests into
a separate base class from the tests.  This allows the same setup to be
reused for tests of the unified search.
